requirements.txt
aiohttp>=3.9.0
slack-sdk==3.27.1
openpyxl>=3.1.0
python-dotenv==1.0.1
pandas==2.2.1
requests==2.31.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0
python-dateutil==2.8.2 